src/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.0
scikit-learn==1.3.0
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
